+++ library/swift/src/BUILD
@@ -8,7 +8,8 @@ swift_static_framework(
         "Client.swift",
         "Envoy.swift",
         "EnvoyBuilder.swift",
-        "Error.swift",
+        "EnvoyError.swift",
+        "EnvoyStreamEmitter.swift",
         "LogLevel.swift",
         "Request.swift",
         "RequestBuilder.swift",
